@@ -1,6 +1,6 @@
 {
   "name": "webchat",
-  "version": "1.4.6",
+  "version": "1.4.7",
   "description": "",
   "main": "lib/index.js",
   "scripts": {
@@ -30,7 +30,7 @@
     "react-dom": "^16.1.1",
     "react-markdown": "^4.0.6",
     "react-redux": "^5.0.6",
-    "react-slick": "^0.18.1",
+    "react-slick": "^0.25.0",
     "redux": "^3.7.2",
     "redux-actions": "^2.2.1",
     "redux-thunk": "^2.2.0",
@@ -38,34 +38,51 @@
     "valid-url": "^1.0.9"
   },
   "devDependencies": {
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-do-expressions": "^7.0.0",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-bind": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime-corejs2": "^7.0.0",
     "autoprefixer": "^7.1.6",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
     "babel-eslint": "^10.0.1",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-runtime": "^6.26.0",
-    "css-loader": "^0.28.7",
-    "eslint": "^4.11.0",
-    "eslint-config-prettier": "^2.7.0",
-    "eslint-config-zavatta": "^6.0.1",
-    "eslint-config-zavatta-react": "^2.3.0",
-    "eslint-plugin-react": "^7.4.0",
+    "babel-loader": "^8.0.0",
+    "core-js": "^2.6.8",
+    "css-loader": "^3.2.0",
+    "eslint": "^5.16.0",
+    "eslint-config-prettier": "^6.1.0",
+    "eslint-config-zavatta": "^6.0.3",
+    "eslint-config-zavatta-react": "^2.3.1",
+    "eslint-plugin-react": "^7.14.3",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.5",
+    "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.6.1",
     "path": "^0.12.7",
-    "postcss-loader": "^2.0.8",
+    "postcss-loader": "^3.0.0",
     "precss": "^2.0.0",
     "prettier": "^1.8.2",
     "progress-bar-webpack-plugin": "^1.10.0",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.19.0",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.4"
+    "sass-loader": "^7.3.1",
+    "style-loader": "^1.0.0",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.41.2",
+    "webpack-dev-server": "^3.9.0"
   }
 }
