@@ -5,9 +5,9 @@ import { store } from 'store'
 import Webchat from './containers/App'
 
 // https://github.com/babel/babel-loader/issues/401
-if (!global._babelPolyfill) {
-  require('babel-polyfill')
-}
+// if (!global._babelPolyfill) {
+//   require('@babel/polyfill')
+// }
 
 export default class CaiWebchat extends Component {
   render () {
