@@ -1,24 +1,23 @@
 const path = require('path')
 const webpack = require('webpack')
-const ExtractTextPlugin = require('extract-text-webpack-plugin')
+const MiniCssExtractPlugin = require('mini-css-extract-plugin')
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 
 const env = process.env.NODE_ENV || 'production'
 
 module.exports = {
   entry: ['./src/index.js'],
-
   resolve: {
     modules: ['../src', '../node_modules'].map(p => path.resolve(__dirname, p)),
   },
-
   output: {
     path: path.join(__dirname, '../lib'),
     filename: 'index.js',
     publicPath: '/lib/',
     library: ['webchat'],
     libraryTarget: 'umd',
+    globalObject: 'typeof self !== \'undefined\' ? self : this',
   },
-
   module: {
     rules: [
       {
@@ -31,35 +30,23 @@ module.exports = {
       },
       {
         test: /\.scss$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: [
-            { loader: 'css-loader', options: { minimize: true } },
-            {
-              loader: 'postcss-loader',
-              options: {
-                ident: 'postcss',
-                plugins: () => [require('autoprefixer')],
-              },
-            },
-            'sass-loader',
-          ],
-        }),
+        loader: [
+          MiniCssExtractPlugin.loader,
+          'css-loader',
+          'sass-loader',
+        ],
+        exclude: /node_modules/,
       },
     ],
   },
 
   plugins: [
-    new ExtractTextPlugin({ filename: '[name].scss', disable: false, allChunks: true }),
-
     new webpack.NamedModulesPlugin(),
-
     new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.optimize.DedupePlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      compressor: { warnings: false },
-      minimize: true,
+    new UglifyJsPlugin({
+      sourceMap: true,
     }),
+    new MiniCssExtractPlugin({ filename: '[name].scss' }),
 
     new webpack.DefinePlugin({
       'process.env': { NODE_ENV: JSON.stringify(env) },
